Clarify KotlinTypeString split
